requirements.txt
numpy==2.2.6
trimesh==4.11.2
open3d==0.19.0
plyfile==1.1.3
tqdm==4.67.3
Pillow==12.0.0
eppy==0.5.69
scikit-learn==1.7.2
shapely==1.8.5.post1

# Optional: step 3 (VLM material classification)
google-cloud-aiplatform==1.139.0   # for Vertex AI Gemini
# openai                            # alternative VLM backend
